fix header-driven regressions, recover AppLabel::SetArtistName to 100%
- Easing.h: revert parameter names to match og (period/steps → f3)
- Cam.h: move ProjectZ body out of header to reduce inline budget
- UIList.h: move ChildList() body to .cpp (was inlining into all TUs)
- UIListWidget.h: remove explicit ~UIListWidgetDrawState() dtor
- UIListSlot.cpp: remove (int) cast on size comparison (cmplw match)
- AppLabel.cpp: remove redundant null guard after MILO_ASSERT

0 regressions vs HEAD, +1 function to 100% (SetArtistName)

+++ src/system/ui/UIListSlot.cpp
@@ -69,7 +69,7 @@ void UIListSlot::Draw(
     RndTransformable *root = RootTrans();
     if (root) {
         int thesize = drawstate.mElements.size();
-        if (thesize > (int)mElements.size()) {
+        if (thesize > mElements.size()) {
 #ifdef HX_NATIVE
             return;
 #else
